requirements.txt
fastapi
uvicorn
torch
pydantic
python-multipart
Pillow
torchvision
